Changed `moisture_read` output. Moved tcp message callback functions to its own file.
- `moisture_read` now returns a percentage from 0 to 100 instead of 1023 to 0 the wetter the environment
- Moved the tcp message related functions into its own files

+++ IoT/lib/drivers/moisture.h
@@ -2,4 +2,4 @@
 #include <stdint.h>
 
 void moisture_init();
-uint16_t moisture_read();
+uint8_t moisture_read();
